Conflicts:
	src/server/game/Entities/Creature/Creature.cpp
	src/server/game/Maps/Map.cpp
	src/server/scripts/EasternKingdoms/SunkenTemple/sunken_temple.h
	src/server/scripts/Northrend/IcecrownCitadel/boss_sindragosa.cpp

+++ src/server/game/Movement/MovementGenerators/HomeMovementGenerator.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2010 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2008-2011 TrinityCore <http://www.trinitycore.org/>
  * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
  *
  * This program is free software; you can redistribute it and/or modify it
@@ -60,7 +60,7 @@ HomeMovementGenerator<Creature>::_setTargetLocation(Creature & owner)
 
     uint32 travel_time = i_destinationHolder.SetDestination(traveller, x, y, z);
     modifyTravelTime(travel_time);
-    owner.ClearUnitState(UNIT_STAT_ALL_STATE & ~UNIT_STAT_EVADE);
+    owner.ClearUnitState(uint32(UNIT_STAT_ALL_STATE & ~UNIT_STAT_EVADE));
 }
 
 bool
